requirements.txt
torch
torchvision
transformers
numpy
pillow
fastapi
uvicorn